pyproject.toml
[project]
name = "codesage"
version = "0.1.0"
description = "A code retrieval system using LlamaIndex and LanceDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-core==0.12.42",
    "llama-index-embeddings-huggingface==0.5.4",
    "llama-index-vector-stores-lancedb==0.3.2",
    "lancedb==0.23.0",
    "transformers==4.52.4",
    "tree-sitter==0.24.0",
    "tree-sitter-languages==1.10.2",
    "pandas==2.3.0",
    "sentence-transformers==4.1.0",
    "torch==2.7.1",
    "numpy==2.3.0",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "safetensors==0.5.3",
    "huggingface-hub==0.33.0",
    "pyarrow==20.0.0",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "sqlalchemy==2.0.41",
    "tantivy==0.24.0",
    "nltk==3.9.1",
    "networkx==3.5",
    "tqdm==4.67.1",
    "requests==2.32.4",
    "httpx==0.28.1",
    "aiohttp==3.12.12",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "tzdata==2025.2",
    "pyyaml==6.0.2",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "packaging==25.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "idna==3.10",
    "urllib3==2.4.0",
    "click==8.2.1",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "dirtyjson==1.0.8",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "fsspec==2025.5.1",
    "greenlet==3.2.3",
    "griffe==1.7.3",
    "h11==0.16.0",
    "hf-xet==1.1.3",
    "httpcore==1.0.9",
    "joblib==1.5.1",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "overrides==7.7.0",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "propcache==0.3.2",
    "pylance==0.29.0",
    "regex==2024.11.6",
    "sniffio==1.3.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "typing-extensions==4.14.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "wrapt==1.17.2",
    "yarl==1.20.1",
    "aiohappyeyeballs==2.6.1",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "banks==2.1.2",
    "frozenlist==1.7.0",
    "multidict==6.4.4",
    "tree-sitter-language-pack>=0.8.0",
    "panda>=0.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
